Move daemonconfig into daemon
