+++ daemon/config.go
@@ -1,4 +1,4 @@
-package daemonconfig
+package daemon
 
 import (
 	"github.com/docker/docker/daemon/networkdriver"
@@ -10,6 +10,9 @@ const (
 	DisableNetworkBridge = "none"
 )
 
+// Config define the configuration of a docker daemon
+//  These are the configuration settings that you pass
+// to the docker daemon when you launch it with say: `docker -d -e lxc`
 // FIXME: separate runtime configuration from http api configuration
 type Config struct {
 	Pidfile                     string
