--- daemonconfig/README.md
@@ -1,3 +1,0 @@
-This directory contains code pertaining to the configuration of the docker daemon
-
-These are the configuration settings that you pass to the docker daemon when you launch it with say: `docker -d -e lxc`
